requirements.txt
seaborn
pillow
numpy